Change file import to use the same code as key import on pre-Kitkat
We have 2 different way of handling import on pre-Kitkat:

1. Use hardcoded file name (cards import)
2. Use an intent but different than post-Kitkat one (key import)

So let's settle on one: an intent and merge the code.

I tested it on my Blackberry Q10 that implements android 4.3 VM.
